% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Blog.posts}
\alias{Blog.posts}
\title{Blog.posts Object}
\usage{
Blog.posts(items = NULL, selfLink = NULL, totalItems = NULL)
}
\arguments{
\item{items}{The List of Posts for this Blog}

\item{selfLink}{The URL of the container for posts in this blog}

\item{totalItems}{The count of posts in this blog}
}
\value{
Blog.posts object
}
\description{
Blog.posts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The container of posts in this blog.
}
\seealso{
Other Blog functions: \code{\link{Blog.locale}},
  \code{\link{Blog.pages}}, \code{\link{Blog}}
}
