% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Blog}
\alias{Blog}
\title{Blogger API Objects 
API for access to the data within Blogger.}
\usage{
Blog(Blog.locale = NULL, Blog.pages = NULL, Blog.posts = NULL,
  customMetaData = NULL, description = NULL, id = NULL, locale = NULL,
  name = NULL, pages = NULL, posts = NULL, published = NULL,
  selfLink = NULL, status = NULL, updated = NULL, url = NULL)
}
\arguments{
\item{Blog.locale}{The \link{Blog.locale} object or list of objects}

\item{Blog.pages}{The \link{Blog.pages} object or list of objects}

\item{Blog.posts}{The \link{Blog.posts} object or list of objects}

\item{customMetaData}{The JSON custom meta-data for the Blog}

\item{description}{The description of this blog}

\item{id}{The identifier for this resource}

\item{locale}{The locale this Blog is set to}

\item{name}{The name of this blog}

\item{pages}{The container of pages in this blog}

\item{posts}{The container of posts in this blog}

\item{published}{RFC 3339 date-time when this blog was published}

\item{selfLink}{The API REST URL to fetch this resource from}

\item{status}{The status of the blog}

\item{updated}{RFC 3339 date-time when this blog was last updated}

\item{url}{The URL where this blog is published}
}
\value{
Blog object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:29:40
filename: /Users/mark/dev/R/autoGoogleAPI/googlebloggerv3.auto/R/blogger_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Blog Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Blog functions: \code{\link{Blog.locale}},
  \code{\link{Blog.pages}}, \code{\link{Blog.posts}}
}
