% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{BlogPerUserInfo}
\alias{BlogPerUserInfo}
\title{BlogPerUserInfo Object}
\usage{
BlogPerUserInfo(blogId = NULL, hasAdminAccess = NULL,
  photosAlbumKey = NULL, role = NULL, userId = NULL)
}
\arguments{
\item{blogId}{ID of the Blog resource}

\item{hasAdminAccess}{True if the user has Admin level access to the blog}

\item{photosAlbumKey}{The Photo Album Key for the user when adding photos to the blog}

\item{role}{Access permissions that the user has for the blog (ADMIN, AUTHOR, or READER)}

\item{userId}{ID of the User}
}
\value{
BlogPerUserInfo object
}
\description{
BlogPerUserInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

