% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{blogs.get}
\alias{blogs.get}
\title{Gets one blog by ID.}
\usage{
blogs.get(blogId, maxPosts = NULL, view = NULL)
}
\arguments{
\item{blogId}{The ID of the blog to get}

\item{maxPosts}{Maximum number of posts to pull back with the blog}

\item{view}{Access level with which to view the blog}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}

