% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{blogs.getByUrl}
\alias{blogs.getByUrl}
\title{Retrieve a Blog by URL.}
\usage{
blogs.getByUrl(url, view = NULL)
}
\arguments{
\item{url}{The URL of the blog to retrieve}

\item{view}{Access level with which to view the blog}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}
