% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{blogs.listByUser}
\alias{blogs.listByUser}
\title{Retrieves a list of blogs, possibly filtered.}
\usage{
blogs.listByUser(userId, fetchUserInfo = NULL, role = NULL, status = NULL,
  view = NULL)
}
\arguments{
\item{userId}{ID of the user whose blogs are to be fetched}

\item{fetchUserInfo}{Whether the response is a list of blogs with per-user information instead of just blogs}

\item{role}{User access types for blogs to include in the results, e}

\item{status}{Blog statuses to include in the result (default: Live blogs only)}

\item{view}{Access level with which to view the blogs}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}
