% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{BlogUserInfo}
\alias{BlogUserInfo}
\title{BlogUserInfo Object}
\usage{
BlogUserInfo(blog = NULL, blog_user_info = NULL)
}
\arguments{
\item{blog}{The Blog resource}

\item{blog_user_info}{Information about a User for the Blog}
}
\value{
BlogUserInfo object
}
\description{
BlogUserInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

