% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Comment.blog}
\alias{Comment.blog}
\title{Comment.blog Object}
\usage{
Comment.blog(id = NULL)
}
\arguments{
\item{id}{The identifier of the blog containing this comment}
}
\value{
Comment.blog object
}
\description{
Comment.blog Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data about the blog containing this comment.
}
\seealso{
Other Comment functions: \code{\link{Comment.author.image}},
  \code{\link{Comment.author}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.post}}, \code{\link{Comment}}
}

