% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Comment.inReplyTo}
\alias{Comment.inReplyTo}
\title{Comment.inReplyTo Object}
\usage{
Comment.inReplyTo(id = NULL)
}
\arguments{
\item{id}{The identified of the parent of this comment}
}
\value{
Comment.inReplyTo object
}
\description{
Comment.inReplyTo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data about the comment this is in reply to.
}
\seealso{
Other Comment functions: \code{\link{Comment.author.image}},
  \code{\link{Comment.author}}, \code{\link{Comment.blog}},
  \code{\link{Comment.post}}, \code{\link{Comment}}
}

