% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Comment}
\alias{Comment}
\title{Comment Object}
\usage{
Comment(Comment.author = NULL, Comment.author.image = NULL,
  Comment.blog = NULL, Comment.inReplyTo = NULL, Comment.post = NULL,
  author = NULL, blog = NULL, content = NULL, id = NULL,
  inReplyTo = NULL, post = NULL, published = NULL, selfLink = NULL,
  status = NULL, updated = NULL)
}
\arguments{
\item{Comment.author}{The \link{Comment.author} object or list of objects}

\item{Comment.author.image}{The \link{Comment.author.image} object or list of objects}

\item{Comment.blog}{The \link{Comment.blog} object or list of objects}

\item{Comment.inReplyTo}{The \link{Comment.inReplyTo} object or list of objects}

\item{Comment.post}{The \link{Comment.post} object or list of objects}

\item{author}{The author of this Comment}

\item{blog}{Data about the blog containing this comment}

\item{content}{The actual content of the comment}

\item{id}{The identifier for this resource}

\item{inReplyTo}{Data about the comment this is in reply to}

\item{post}{Data about the post containing this comment}

\item{published}{RFC 3339 date-time when this comment was published}

\item{selfLink}{The API REST URL to fetch this resource from}

\item{status}{The status of the comment (only populated for admin users)}

\item{updated}{RFC 3339 date-time when this comment was last updated}
}
\value{
Comment object
}
\description{
Comment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Comment functions: \code{\link{Comment.author.image}},
  \code{\link{Comment.author}}, \code{\link{Comment.blog}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.post}}
}
