% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{comments.delete}
\alias{comments.delete}
\title{Delete a comment by ID.}
\usage{
comments.delete(blogId, postId, commentId)
}
\arguments{
\item{blogId}{The ID of the Blog}

\item{postId}{The ID of the Post}

\item{commentId}{The ID of the comment to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}

