% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{comments.get}
\alias{comments.get}
\title{Gets one comment by ID.}
\usage{
comments.get(blogId, postId, commentId, view = NULL)
}
\arguments{
\item{blogId}{ID of the blog to containing the comment}

\item{postId}{ID of the post to fetch posts from}

\item{commentId}{The ID of the comment to get}

\item{view}{Access level for the requested comment (default: READER)}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}
