% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{comments.list}
\alias{comments.list}
\title{Retrieves the comments for a post, possibly filtered.}
\usage{
comments.list(blogId, postId, endDate = NULL, fetchBodies = NULL,
  maxResults = NULL, pageToken = NULL, startDate = NULL, status = NULL,
  view = NULL)
}
\arguments{
\item{blogId}{ID of the blog to fetch comments from}

\item{postId}{ID of the post to fetch posts from}

\item{endDate}{Latest date of comment to fetch, a date-time with RFC 3339 formatting}

\item{fetchBodies}{Whether the body content of the comments is included}

\item{maxResults}{Maximum number of comments to include in the result}

\item{pageToken}{Continuation token if request is paged}

\item{startDate}{Earliest date of comment to fetch, a date-time with RFC 3339 formatting}

\item{status}{}

\item{view}{Access level with which to view the returned result}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}
