% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Page.author.image}
\alias{Page.author.image}
\title{Page.author.image Object}
\usage{
Page.author.image(url = NULL)
}
\arguments{
\item{url}{The page author's avatar URL}
}
\value{
Page.author.image object
}
\description{
Page.author.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The page author's avatar.
}
\seealso{
Other Page functions: \code{\link{Page.author}},
  \code{\link{Page.blog}}, \code{\link{Page}},
  \code{\link{pages.insert}}, \code{\link{pages.patch}},
  \code{\link{pages.update}}
}

