% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Page.author}
\alias{Page.author}
\title{Page.author Object}
\usage{
Page.author(Page.author.image = NULL, displayName = NULL, id = NULL,
  image = NULL, url = NULL)
}
\arguments{
\item{Page.author.image}{The \link{Page.author.image} object or list of objects}

\item{displayName}{The display name}

\item{id}{The identifier of the Page creator}

\item{image}{The page author's avatar}

\item{url}{The URL of the Page creator's Profile page}
}
\value{
Page.author object
}
\description{
Page.author Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The author of this Page.
}
\seealso{
Other Page functions: \code{\link{Page.author.image}},
  \code{\link{Page.blog}}, \code{\link{Page}},
  \code{\link{pages.insert}}, \code{\link{pages.patch}},
  \code{\link{pages.update}}
}

