% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Page}
\alias{Page}
\title{Page Object}
\usage{
Page(Page.author = NULL, Page.author.image = NULL, Page.blog = NULL,
  author = NULL, blog = NULL, content = NULL, etag = NULL, id = NULL,
  published = NULL, selfLink = NULL, status = NULL, title = NULL,
  updated = NULL, url = NULL)
}
\arguments{
\item{Page.author}{The \link{Page.author} object or list of objects}

\item{Page.author.image}{The \link{Page.author.image} object or list of objects}

\item{Page.blog}{The \link{Page.blog} object or list of objects}

\item{author}{The author of this Page}

\item{blog}{Data about the blog containing this Page}

\item{content}{The body content of this Page, in HTML}

\item{etag}{Etag of the resource}

\item{id}{The identifier for this resource}

\item{published}{RFC 3339 date-time when this Page was published}

\item{selfLink}{The API REST URL to fetch this resource from}

\item{status}{The status of the page for admin resources (either LIVE or DRAFT)}

\item{title}{The title of this entity}

\item{updated}{RFC 3339 date-time when this Page was last updated}

\item{url}{The URL that this Page is displayed at}
}
\value{
Page object
}
\description{
Page Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Page functions: \code{\link{Page.author.image}},
  \code{\link{Page.author}}, \code{\link{Page.blog}},
  \code{\link{pages.insert}}, \code{\link{pages.patch}},
  \code{\link{pages.update}}
}
