% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{PageList}
\alias{PageList}
\title{PageList Object}
\usage{
PageList(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{Etag of the response}

\item{items}{The list of Pages for a Blog}

\item{nextPageToken}{Pagination token to fetch the next page, if one exists}
}
\value{
PageList object
}
\description{
PageList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

