% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{pages.list}
\alias{pages.list}
\title{Retrieves the pages for a blog, optionally including non-LIVE statuses.}
\usage{
pages.list(blogId, fetchBodies = NULL, maxResults = NULL,
  pageToken = NULL, status = NULL, view = NULL)
}
\arguments{
\item{blogId}{ID of the blog to fetch Pages from}

\item{fetchBodies}{Whether to retrieve the Page bodies}

\item{maxResults}{Maximum number of Pages to fetch}

\item{pageToken}{Continuation token if the request is paged}

\item{status}{}

\item{view}{Access level with which to view the returned result}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}
