% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{pages.update}
\alias{pages.update}
\title{Update a page.}
\usage{
pages.update(Page, blogId, pageId, publish = NULL, revert = NULL)
}
\arguments{
\item{Page}{The \link{Page} object to pass to this method}

\item{blogId}{The ID of the Blog}

\item{pageId}{The ID of the Page}

\item{publish}{Whether a publish action should be performed when the page is updated (default: false)}

\item{revert}{Whether a revert action should be performed when the page is updated (default: false)}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}

Other Page functions: \code{\link{Page.author.image}},
  \code{\link{Page.author}}, \code{\link{Page.blog}},
  \code{\link{Page}}, \code{\link{pages.insert}},
  \code{\link{pages.patch}}
}

