% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Pageviews.counts}
\alias{Pageviews.counts}
\title{Pageviews.counts Object}
\usage{
Pageviews.counts()
}
\value{
Pageviews.counts object
}
\description{
Pageviews.counts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Pageviews functions: \code{\link{Pageviews}}
}

