% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Pageviews}
\alias{Pageviews}
\title{Pageviews Object}
\usage{
Pageviews(Pageviews.counts = NULL, blogId = NULL, counts = NULL)
}
\arguments{
\item{Pageviews.counts}{The \link{Pageviews.counts} object or list of objects}

\item{blogId}{Blog Id}

\item{counts}{The container of posts in this blog}
}
\value{
Pageviews object
}
\description{
Pageviews Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Pageviews functions: \code{\link{Pageviews.counts}}
}
