% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Post.author.image}
\alias{Post.author.image}
\title{Post.author.image Object}
\usage{
Post.author.image(url = NULL)
}
\arguments{
\item{url}{The Post author's avatar URL}
}
\value{
Post.author.image object
}
\description{
Post.author.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Post author's avatar.
}
\seealso{
Other Post functions: \code{\link{Post.author}},
  \code{\link{Post.blog}}, \code{\link{Post.images}},
  \code{\link{Post.location}}, \code{\link{Post.replies}},
  \code{\link{Post}}, \code{\link{posts.insert}},
  \code{\link{posts.patch}}, \code{\link{posts.update}}
}
