% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Post.blog}
\alias{Post.blog}
\title{Post.blog Object}
\usage{
Post.blog(id = NULL)
}
\arguments{
\item{id}{The identifier of the Blog that contains this Post}
}
\value{
Post.blog object
}
\description{
Post.blog Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data about the blog containing this Post.
}
\seealso{
Other Post functions: \code{\link{Post.author.image}},
  \code{\link{Post.author}}, \code{\link{Post.images}},
  \code{\link{Post.location}}, \code{\link{Post.replies}},
  \code{\link{Post}}, \code{\link{posts.insert}},
  \code{\link{posts.patch}}, \code{\link{posts.update}}
}
