% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Post.location}
\alias{Post.location}
\title{Post.location Object}
\usage{
Post.location(lat = NULL, lng = NULL, name = NULL, span = NULL)
}
\arguments{
\item{lat}{Location's latitude}

\item{lng}{Location's longitude}

\item{name}{Location name}

\item{span}{Location's viewport span}
}
\value{
Post.location object
}
\description{
Post.location Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The location for geotagged posts.
}
\seealso{
Other Post functions: \code{\link{Post.author.image}},
  \code{\link{Post.author}}, \code{\link{Post.blog}},
  \code{\link{Post.images}}, \code{\link{Post.replies}},
  \code{\link{Post}}, \code{\link{posts.insert}},
  \code{\link{posts.patch}}, \code{\link{posts.update}}
}
