% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{PostList}
\alias{PostList}
\title{PostList Object}
\usage{
PostList(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{Etag of the response}

\item{items}{The list of Posts for this Blog}

\item{nextPageToken}{Pagination token to fetch the next page, if one exists}
}
\value{
PostList object
}
\description{
PostList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

