% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{PostPerUserInfo}
\alias{PostPerUserInfo}
\title{PostPerUserInfo Object}
\usage{
PostPerUserInfo(blogId = NULL, hasEditAccess = NULL, postId = NULL,
  userId = NULL)
}
\arguments{
\item{blogId}{ID of the Blog that the post resource belongs to}

\item{hasEditAccess}{True if the user has Author level access to the post}

\item{postId}{ID of the Post resource}

\item{userId}{ID of the User}
}
\value{
PostPerUserInfo object
}
\description{
PostPerUserInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
