% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{posts.get}
\alias{posts.get}
\title{Get a post by ID.}
\usage{
posts.get(blogId, postId, fetchBody = NULL, fetchImages = NULL,
  maxComments = NULL, view = NULL)
}
\arguments{
\item{blogId}{ID of the blog to fetch the post from}

\item{postId}{The ID of the post}

\item{fetchBody}{Whether the body content of the post is included (default: true)}

\item{fetchImages}{Whether image URL metadata for each post is included (default: false)}

\item{maxComments}{Maximum number of comments to pull back on a post}

\item{view}{Access level with which to view the returned result}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}

