% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{posts.insert}
\alias{posts.insert}
\title{Add a post.}
\usage{
posts.insert(Post, blogId, fetchBody = NULL, fetchImages = NULL,
  isDraft = NULL)
}
\arguments{
\item{Post}{The \link{Post} object to pass to this method}

\item{blogId}{ID of the blog to add the post to}

\item{fetchBody}{Whether the body content of the post is included with the result (default: true)}

\item{fetchImages}{Whether image URL metadata for each post is included in the returned result (default: false)}

\item{isDraft}{Whether to create the post as a draft (default: false)}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}

Other Post functions: \code{\link{Post.author.image}},
  \code{\link{Post.author}}, \code{\link{Post.blog}},
  \code{\link{Post.images}}, \code{\link{Post.location}},
  \code{\link{Post.replies}}, \code{\link{Post}},
  \code{\link{posts.patch}}, \code{\link{posts.update}}
}
