% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{posts.list}
\alias{posts.list}
\title{Retrieves a list of posts, possibly filtered.}
\usage{
posts.list(blogId, endDate = NULL, fetchBodies = NULL, fetchImages = NULL,
  labels = NULL, maxResults = NULL, orderBy = NULL, pageToken = NULL,
  startDate = NULL, status = NULL, view = NULL)
}
\arguments{
\item{blogId}{ID of the blog to fetch posts from}

\item{endDate}{Latest post date to fetch, a date-time with RFC 3339 formatting}

\item{fetchBodies}{Whether the body content of posts is included (default: true)}

\item{fetchImages}{Whether image URL metadata for each post is included}

\item{labels}{Comma-separated list of labels to search for}

\item{maxResults}{Maximum number of posts to fetch}

\item{orderBy}{Sort search results}

\item{pageToken}{Continuation token if the request is paged}

\item{startDate}{Earliest post date to fetch, a date-time with RFC 3339 formatting}

\item{status}{Statuses to include in the results}

\item{view}{Access level with which to view the returned result}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}
