% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{posts.patch}
\alias{posts.patch}
\title{Update a post. This method supports patch semantics.}
\usage{
posts.patch(Post, blogId, postId, fetchBody = NULL, fetchImages = NULL,
  maxComments = NULL, publish = NULL, revert = NULL)
}
\arguments{
\item{Post}{The \link{Post} object to pass to this method}

\item{blogId}{The ID of the Blog}

\item{postId}{The ID of the Post}

\item{fetchBody}{Whether the body content of the post is included with the result (default: true)}

\item{fetchImages}{Whether image URL metadata for each post is included in the returned result (default: false)}

\item{maxComments}{Maximum number of comments to retrieve with the returned post}

\item{publish}{Whether a publish action should be performed when the post is updated (default: false)}

\item{revert}{Whether a revert action should be performed when the post is updated (default: false)}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}

Other Post functions: \code{\link{Post.author.image}},
  \code{\link{Post.author}}, \code{\link{Post.blog}},
  \code{\link{Post.images}}, \code{\link{Post.location}},
  \code{\link{Post.replies}}, \code{\link{Post}},
  \code{\link{posts.insert}}, \code{\link{posts.update}}
}
