% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{posts.publish}
\alias{posts.publish}
\title{Publishes a draft post, optionally at the specific time of the given publishDate parameter.}
\usage{
posts.publish(blogId, postId, publishDate = NULL)
}
\arguments{
\item{blogId}{The ID of the Blog}

\item{postId}{The ID of the Post}

\item{publishDate}{Optional date and time to schedule the publishing of the Blog}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}

