% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{posts.search}
\alias{posts.search}
\title{Search for a post.}
\usage{
posts.search(blogId, q, fetchBodies = NULL, orderBy = NULL)
}
\arguments{
\item{blogId}{ID of the blog to fetch the post from}

\item{q}{Query terms to search this blog for matching posts}

\item{fetchBodies}{Whether the body content of posts is included (default: true)}

\item{orderBy}{Sort search results}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}

