% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{PostUserInfo}
\alias{PostUserInfo}
\title{PostUserInfo Object}
\usage{
PostUserInfo(post = NULL, post_user_info = NULL)
}
\arguments{
\item{post}{The Post resource}

\item{post_user_info}{Information about a User for the Post}
}
\value{
PostUserInfo object
}
\description{
PostUserInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
