% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{postUserInfos.get}
\alias{postUserInfos.get}
\title{Gets one post and user info pair, by post ID and user ID. The post user info contains per-user information about the post, such as access rights, specific to the user.}
\usage{
postUserInfos.get(userId, blogId, postId, maxComments = NULL)
}
\arguments{
\item{userId}{ID of the user for the per-user information to be fetched}

\item{blogId}{The ID of the blog}

\item{postId}{The ID of the post to get}

\item{maxComments}{Maximum number of comments to pull back on a post}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
\item https://www.googleapis.com/auth/blogger.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger, https://www.googleapis.com/auth/blogger.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/3.0/getting_started}{Google Documentation}
}
