% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{PostUserInfosList}
\alias{PostUserInfosList}
\title{PostUserInfosList Object}
\usage{
PostUserInfosList(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The list of Posts with User information for the post, for this Blog}

\item{nextPageToken}{Pagination token to fetch the next page, if one exists}
}
\value{
PostUserInfosList object
}
\description{
PostUserInfosList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

