% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{User.locale}
\alias{User.locale}
\title{User.locale Object}
\usage{
User.locale(country = NULL, language = NULL, variant = NULL)
}
\arguments{
\item{country}{The user's country setting}

\item{language}{The user's language setting}

\item{variant}{The user's language variant setting}
}
\value{
User.locale object
}
\description{
User.locale Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This user's locale
}
\seealso{
Other User functions: \code{\link{User.blogs}},
  \code{\link{User}}
}

